requirements.txt
django
djangorestframework
git+https://github.com/daniel-waruo/django-rest-auth.git@master#egg=django-rest-auth
django-allauth

pillow
django-qsessions

django-rest-knox
pyuploadcare

python-dotenv
django-split-settings

gunicorn
django_heroku

django-cors-headers

graphene_django
django-filter

africastalking